Split Storage module into storage_xml and storage_json to enable credentials-based switching between them
