+++ lib/fog/google.rb
@@ -21,7 +21,7 @@ module Fog
     autoload :SQL, "fog/google/sql"
 
     # TODO: Deprecate old storage API, make JSON main API.
-    autoload :StorageJSON, "fog/google/storage_json"
+    # autoload :StorageJSON, "fog/google/storage_json"
 
     extend Fog::Provider
 
@@ -29,7 +29,6 @@ module Fog
     service(:dns, "DNS")
     service(:monitoring, "Monitoring")
     service(:storage, "Storage")
-    service(:storage_json, "StorageJSON")
     service(:sql, "SQL")
 
     # CGI.escape, but without special treatment on spaces
